refactor: Payment에서 Command, Query Controller 분리
