+++ src/main/java/cotw/server/domain/payment/controller/PaymentController.java
@@ -6,14 +6,10 @@ import cotw.server.domain.payment.dto.request.PaymentConfirmRequest;
 import cotw.server.domain.payment.dto.request.PaymentCreateRequest;
 import cotw.server.domain.payment.dto.response.PaymentCancelResponse;
 import cotw.server.domain.payment.dto.response.PaymentCreateResponse;
-import cotw.server.domain.payment.dto.response.PaymentHistoryResponse;
 import cotw.server.domain.payment.service.PaymentService;
-import cotw.server.domain.payment.service.LedgerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/payments")
@@ -21,7 +17,6 @@ import java.util.List;
 public class PaymentController {
 
     private final PaymentService paymentService;
-    private final LedgerService ledgerService;
 
     @PostMapping
     public ResponseEntity<PaymentCreateResponse> createPayment(@RequestBody PaymentCreateRequest request) {
@@ -52,25 +47,6 @@ public class PaymentController {
                     .header("Location", "http://localhost:5173/payment/fail?error=payment_failed")
                     .build();
         }
-    }
-
-    @GetMapping("/member/{memberId}")
-    public ResponseEntity<List<PaymentHistoryResponse>> getPaymentsByMember(@PathVariable Long memberId) {
-        List<PaymentHistoryResponse> responses = ledgerService.getPaymentHistoryByMember(memberId);
-        return ResponseEntity.ok(responses);
-    }
-
-    @GetMapping("/post/{postId}")
-    public ResponseEntity<List<PaymentHistoryResponse>> getPaymentsByPost(@PathVariable Long postId) {
-        List<PaymentHistoryResponse> responses = ledgerService.getPaymentHistoryByPost(postId);
-        return ResponseEntity.ok(responses);
-    }
-
-    @GetMapping("/my")
-    public ResponseEntity<List<PaymentHistoryResponse>> getMyPayments() {
-        Long memberId = SecurityUtil.getCurrentMemberId();
-        List<PaymentHistoryResponse> responses = ledgerService.getPaymentHistoryByMember(memberId);
-        return ResponseEntity.ok(responses);
     }
 
     @PostMapping("/cancel")
